Add files via upload
In the real world, you will often need to handle missing values. One way to impute (i.e., fill) the numerical column is to replace the null values with their mean.
Task
Given a list of numbers including some missing values, turn it into a pandas series, impute the missing values with the mean, and finally return the series.